+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## 4.9.2 (unreleased)
 
 
-- Nothing changed yet.
+- Support extract strategies
 
 
 ## 4.9.1 (2025-06-10)